// src/types/treasuryDirect.ts
export interface Security {
    cusip: string
    issueDate: string
    securityType: string
    securityTerm: string
    maturityDate: string
    interestRate: string
    refCpiOnIssueDate: string
    refCpiOnDatedDate: string
    announcementDate: string
    auctionDate: string
    auctionDateYear: string
    datedDate: string
    accruedInterestPer1000: string
    accruedInterestPer100: string
    adjustedAccruedInterestPer1000: string
    adjustedPrice: string
    allocationPercentage: string
    allocationPercentageDecimals: string
    announcedCusip: string
    auctionFormat: string
    averageMedianDiscountRate: string
    averageMedianInvestmentRate: string
    averageMedianPrice: string
    averageMedianDiscountMargin: string
    averageMedianYield: string
    backDated: string
    backDatedDate: string
    bidToCoverRatio: string
    callDate: string
    callable: string
    calledDate: string
    cashManagementBillCMB: string
    closingTimeCompetitive: string
    closingTimeNoncompetitive: string
    competitiveAccepted: string
    competitiveBidDecimals: string
    competitiveTendered: string
    competitiveTendersAccepted: string
    corpusCusip: string
    cpiBaseReferencePeriod: string
    currentlyOutstanding: string
    directBidderAccepted: string
    directBidderTendered: string
    estimatedAmountOfPubliclyHeldMaturingSecuritiesByType: string
    fimaIncluded: string
    fimaNoncompetitiveAccepted: string
    fimaNoncompetitiveTendered: string
    firstInterestPeriod: string
    firstInterestPaymentDate: string
    floatingRate: string
    frnIndexDeterminationDate: string
    frnIndexDeterminationRate: string
    highDiscountRate: string
    highInvestmentRate: string
    highPrice: string
    highDiscountMargin: string
    highYield: string
    indexRatioOnIssueDate: string
    indirectBidderAccepted: string
    indirectBidderTendered: string
    interestPaymentFrequency: string
    lowDiscountRate: string
    lowInvestmentRate: string
    lowPrice: string
    lowDiscountMargin: string
    lowYield: string
    maturingDate: string
    maximumCompetitiveAward: string
    maximumNoncompetitiveAward: string
    maximumSingleBid: string
    minimumBidAmount: string
    minimumStripAmount: string
    minimumToIssue: string
    multiplesToBid: string
    multiplesToIssue: string
    nlpExclusionAmount: string
    nlpReportingThreshold: string
    noncompetitiveAccepted: string
    noncompetitiveTendersAccepted: string
    offeringAmount: string
    originalCusip: string
    originalDatedDate: string
    originalIssueDate: string
    originalSecurityTerm: string
    pdfFilenameAnnouncement: string
    pdfFilenameCompetitiveResults: string
    pdfFilenameNoncompetitiveResults: string
    pdfFilenameSpecialAnnouncement: string
    pricePer100: string
    primaryDealerAccepted: string
    primaryDealerTendered: string
    reopening: string
    securityTermDayMonth: string
    securityTermWeekYear: string
    series: string
    somaAccepted: string
    somaHoldings: string
    somaIncluded: string
    somaTendered: string
    spread: string
    standardInterestPaymentPer1000: string
    strippable: string
    term: string
    tiinConversionFactorPer1000: string
    tips: string
    totalAccepted: string
    totalTendered: string
    treasuryRetailAccepted: string
    treasuryRetailTendersAccepted: string
    type: string
    unadjustedAccruedInterestPer1000: string
    unadjustedPrice: string
    updatedTimestamp: string
    xmlFilenameAnnouncement: string
    xmlFilenameCompetitiveResults: string
    xmlFilenameSpecialAnnouncement: string
    tintCusip1: string
    tintCusip2: string
}

export type SECURITY_TYPES_TYPE = 'Bill' | 'Note' | 'Bond' | 'CMB' | 'TIPS' | 'FRN'